progress of adding dll decls for funcs without types
