+++ tsc/tsc-new-parser/config.h
@@ -22,6 +22,7 @@ using regex = std::wregex;
 using sregex_iterator = std::wsregex_iterator;
 using stringstream = std::wstringstream;
 using smatch = std::wsmatch;
+using ostream = std::wostream;
 #define regex_replace std::regex_replace
 
 #define to_number_base(x, y) std::stoi(x, nullptr, y)
